requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
cfgv==3.4.0
charset-normalizer==3.3.2
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.15.4
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.5.36
idna==3.7
ipykernel==6.29.4
ipython==8.26.0
isort==5.13.2
jedi==0.19.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-openai==0.1.14
langchain-text-splitters==0.2.2
langsmith==0.1.82
marshmallow==3.21.3
matplotlib-inline==0.1.7
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==1.26.4
openai==1.35.9
orjson==3.10.5
packaging==24.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
pre-commit==3.7.1
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
ruff==0.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
tenacity==8.4.2
tiktoken==0.7.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
wcwidth==0.2.13
yarl==1.9.4
